fix aligning of team and pagination
